Use BottomSheetScaffold
